+++ library/src/schemas/nonNullish/index.ts
@@ -1,2 +1,3 @@
 export * from './nonNullish.ts';
 export * from './nonNullishAsync.ts';
+export * from './types.ts';